Refactor and update codebase for improved readability and performance
The code has been significantly refactored to improve readability and enhance performance. Essential changes include updates to class constructors, amendments to function signatures, and removal of redundant comments. This update also modifies the logical error in the configuration file concerning address size checking, and changes the 'port' data type from size_t to uint16_t in the 'Configuration' class to be more accurate.

+++ Source_code/Includes/3-configuration.hpp
@@ -52,7 +52,7 @@ struct Server {
   vecStr names;
   bool defaultServer;
   string host;
-  size_t port;
+  uint16_t port;
   string root;
   mapIntStr errorPages;
   size_t clientMaxBodySize;
